Add SequentialExecutor class
The goal of this class is to factorize a bit of logic we have everywhere
in tsrc code: running something on several repositories, and collect
error in the end.

For instance, if you run `tsrc init` and one of the repositories could
not be cloned because of a transient network error, you still want
the rest of the repositories to be cloned.