fix when default branch is not master on a given repo
tsrc assumes the default branch of one repository is 'master'. Most of time
this is true, but if the default branch is different (e.g. 'production'), tsrc
may fail. With this commit, there is no default branch set by default, and if
one wants to track one given branch it will have to specified