Refactor: run_git() now takes an explicit `capture` argument.
We no longer assume that calling run_git with raises=False implies
capturing the output, or that capturing the output means we have to
explicitly check the return code.

This has several benefits:

First, we can get rid of:

    rc, out = run_git(..., raises=True)
    if rc != 0:
        raise GitCommandError(...)

and replace that with:

    rc, out = run_git(..., capture=True)
    if rc != 0:
        raise GitCommandError(...)

If we want to capture the output *and* ignore failures, we now hawe to
specify both arguments:

    rc, out = run_git(..., capture=True, raises=False):
    if rc != 0:
        ....

This refactoring also fixes a bug:

If you were on a branch with no upstream set, `tsrc status` used to
display +1 / -1 because it was counting the number of lines of the `git
rev-parse` call, which happens to be of line 1 in case an error occurs.