Modified schema for symlink section in manifest file
The symlink section is now expected to be a peer to copy.  Each
individual symlink needs to specify source and target names for the
desired link.

Also, with the syntax change, fixed the potentially confusing
re-ordering of arguments to os.symlink() within linker.  Everything
should now align reasonably well to os.symlink() documentation and
Unix man pages for ln.