All of tsrc actions can now be run in parallel
How do we do it?

The easy part is to have to classes responsible for executing the same
task over a list of items. Those are SequentialExecutor and
ParallelExecutor

The hard part is to keep tsrc output clean!

In order to do that, we :
 * hide as much as output possible when using parallel tasks
 * introduce a concept of "summary" which can be printed when
   all the items have been processed

More specifically:

Each task now has:

 * A describe_process_start / describe_process_end abstract method which
   is used to keep the parallel output neatly one one line (like ninja-build)
 * Methods like `self.info_3` which can be used to:
    - display messages when sequentially
    - not display anything when run in parallel
 * And the process method now returns an Outcome class which can contain
   a summary or an error message

In order to keep error handling and display of summary easy, we also
introduce a OutcomeCollection class which can display collected errors
and summary

As a neat side-effect, there was some ad-hoc code in `sync`, `foreach` and
`log` which is now just a generic usage of the Outcome class.