tsrc init: Fix order of operations
Following the move from `master` to `main` as a default branch, it is
no longer a good idea to have `tsrc init` use `master` as an hard-coded
default.

Instead, if no branch is specified an the command line, we need to run
`git clone <manifest_url>` without any `--branch` argument.

This mean we can't know in advance which manifest branch is going to be used
(it depends on the *remote head* of the manifest repo).

So we need to clone the local manifest before writing the workspace
config, and thus before instantiating a Workspace() instance.

The good news is that we can simplify LocalManifest code.